+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unpublished
 
-- ...
+- Support for Varnish v6.0 LTS, but without filters, and without `vmod_be`, `vmod_vfp`, `vmod_vdp` examples
+- Require all user functions that return `VCL_*` types to be `unsafe` 
 
 # 0.2.0 (2024-11-19)